Change serializer to use Vec instead of BTreeMap
If we were making large maps and regularly inserting/removing, it'd make
sense to use a BTreeMap. But we're not. We're building up a list of
things, then iterating over a sorted version of it exactly once. A Vec
that gets sorted once and checked for duplicates right after the sort
should be a bit faster than making a map.